+++ articles/views/__init__.py
@@ -1,2 +1,4 @@
 from .detail import ArticleDetailView, CommentsView, FavoriteView, ResponseToView
 from .lists import LatestArticlesView, AuthorArticlesView, CategoryArticlesView
+from .categories import CategoriesListView
+from .usershandler import UserListView